Adjust JournalEntryService function signatures for beer thread safety

+++ VENI VIDI/VENI VIDI/Data/JournalEntryService.swift
@@ -66,7 +66,7 @@ extension JournalEntryService {
 // MARK: - Journal Entry Handling
 
 extension JournalEntryService {
-    func fetchJournalEntries() -> [JournalEntry]? {
+    func fetchAllJournalEntries() -> [JournalEntry]? {
         do {
             let fetchRequest = NSFetchRequest<JournalEntry>(entityName: "JournalEntry")
             let entries = try managedObjectContext.fetch(fetchRequest)
@@ -98,7 +98,7 @@ extension JournalEntryService {
                             atLongitude longitude: Double = 0,
                             atLatitude latitude: Double = 0,
                             withTags tags: [Tag]? = [],
-                            isFavorite favorite: Bool? = false) -> JournalEntry
+                            isFavorite favorite: Bool? = false) -> JournalEntry?
     {
         let newJournalEntry = NSEntityDescription.insertNewObject(forEntityName: "JournalEntry", into: self.managedObjectContext) as! JournalEntry
         // Solution by https://stackoverflow.com/questions/60228931/no-nsentitydescriptions-in-any-model-claim-the-nsmanagedobject-subclass-priorit
@@ -106,22 +106,26 @@ extension JournalEntryService {
         
         newJournalEntry.id = UUID()
         self.coreDataStack.saveContext()
-        
-        self.updateJournalEntry(newJournalEntry,
-                                aboutWork: work,
-                                withCoverImage: coverImage,
-                                withStartDate: startDate,
-                                withFinishDate: finishDate,
-                                withEntryTitle: entryTitle,
-                                withEntryContent: entryContent,
-                                atLongitude: longitude,
-                                atLatitude: latitude,
-                                withTags: tags,
-                                isFavorite: favorite)
-        return newJournalEntry
+        if let entryID = newJournalEntry.id {
+            self.updateJournalEntry(withUUID: entryID,
+                                    aboutWork: work,
+                                    withCoverImage: coverImage,
+                                    withStartDate: startDate,
+                                    withFinishDate: finishDate,
+                                    withEntryTitle: entryTitle,
+                                    withEntryContent: entryContent,
+                                    atLongitude: longitude,
+                                    atLatitude: latitude,
+                                    withTags: tags,
+                                    isFavorite: favorite)
+            return newJournalEntry
+        } else {
+            print("Unable to create JournalEntry: UUID remains nil after assignment")
+            return nil
+        }
     }
     
-    func updateJournalEntry(_ entry: JournalEntry,
+    func updateJournalEntry(withUUID id: UUID,
                             aboutWork work: String? = nil,
                             withCoverImage coverImage: UIImage? = nil,
                             withStartDate startDate: Date? = nil,
@@ -133,43 +137,51 @@ extension JournalEntryService {
                             withTags tags: [Tag]? = nil,
                             isFavorite favorite: Bool? = nil)
     {
-        entry.lastEditDate = Date()
-        entry.worksTitle = work ?? entry.worksTitle
-        entry.startDate = startDate ?? entry.startDate
-        entry.finishDate = finishDate ?? entry.finishDate
-        entry.entryTitle = entryTitle ?? entry.entryTitle
-        entry.entryContent = entryContent ?? entry.entryContent
-        entry.favorite = favorite ?? entry.favorite
+        if let entry = self.fetchJournalEntryWithUUID(id) {
+            entry.lastEditDate = Date()
+            entry.worksTitle = work ?? entry.worksTitle
+            entry.startDate = startDate ?? entry.startDate
+            entry.finishDate = finishDate ?? entry.finishDate
+            entry.entryTitle = entryTitle ?? entry.entryTitle
+            entry.entryContent = entryContent ?? entry.entryContent
+            entry.favorite = favorite ?? entry.favorite
         
-        if let newLongitude = longitude {
-            if let newLatitude = latitude {
-                entry.longitude = newLongitude
-                entry.latitude = newLatitude
-            } else {
-                print("Entry location not updated. Both longitude and latitude needed for update.")
+            if let newLongitude = longitude {
+                if let newLatitude = latitude {
+                    entry.longitude = newLongitude
+                    entry.latitude = newLatitude
+                } else {
+                    print("Entry location not updated. Both longitude and latitude needed for update.")
+                }
             }
-        }
         
-        if let newTags = tags {
-            let oldTags = entry.tags
-            entry.removeFromTags(oldTags ?? NSSet())
-            entry.addToTags(NSSet(array: newTags))
-        }
-        
-        if let newImage = coverImage {
-            // https://stackoverflow.com/questions/16685812/how-to-store-an-image-in-core-data#16687218
-            if let imageData = newImage.pngData() {
-                entry.image = imageData
-            } else {
-                print("Failed to store image in CoreData")
+            if let newTags = tags {
+                let oldTags = entry.tags
+                entry.removeFromTags(oldTags ?? NSSet())
+                entry.addToTags(NSSet(array: newTags))
             }
+        
+            if let newImage = coverImage {
+                // https://stackoverflow.com/questions/16685812/how-to-store-an-image-in-core-data#16687218
+                if let imageData = newImage.pngData() {
+                    entry.image = imageData
+                } else {
+                    print("Failed to store image in CoreData")
+                }
+            }
+        } else {
+            print("Received invalid UUID for updateJournalEntry()")
         }
         
         self.coreDataStack.saveContext()
     }
     
-    func deleteJournalEntry(_ entry: JournalEntry) {
-        self.managedObjectContext.delete(entry)
-        self.coreDataStack.saveContext()
+    func deleteJournalEntry(withUUID id: UUID) {
+        if let entry = self.fetchJournalEntryWithUUID(id) {
+            self.managedObjectContext.delete(entry)
+            self.coreDataStack.saveContext()
+        } else {
+            print("Received invalid UUID for deleteJournalEntry()")
+        }
     }
 }
